long

+++ app/src/main/java/io/husayn/paging_library_sample/listing/PagingRequest.java
@@ -2,11 +2,11 @@ package io.husayn.paging_library_sample.listing;
 
 class PagingRequest {
 
-  public final int offSet;
+  public final long offSet;
   public final PagingQuery pagingQuery;
   public final PagingQueryConfig queryConfig;
 
-  public PagingRequest(int offSet, PagingQuery pagingQuery, PagingQueryConfig queryConfig) {
+  public PagingRequest(long offSet, PagingQuery pagingQuery, PagingQueryConfig queryConfig) {
     this.pagingQuery = pagingQuery;
     this.queryConfig = queryConfig;
     this.offSet = offSet;
